requirements.txt
# Airflow providers
apache-airflow-providers-google==18.0.0
apache-airflow-providers-postgres==6.3.0
apache-airflow-providers-mysql==6.3.4

# Data processing
numpy==1.26.4
pyarrow==18.1.0
pytz==2025.2
polars==1.35.2

# Google Cloud SDKs
google-cloud-bigquery==3.38.0
google-cloud-storage==2.19.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-api-python-client==2.185.0

# DB drivers
PyMySQL==1.1.2
psycopg==3.2.3

# Utilities
PyYAML==6.0.3
attrs==25.4.0
requests==2.32.5
setproctitle==1.3.7




